routing and usersPage refactoring
