+++ src/api/instanceApi.ts
@@ -1,9 +1,9 @@
-import { REFRESH_TOKEN } from "@/components/constants/localStorageValues";
-import { authSlice } from "@/pages/AuthPage/AuthSlice";
 import axios, { AxiosError } from "axios";
+import { REFRESH_TOKEN } from "@/components/constants/localStorageValues";
+import { authSlice } from "../slices/AuthSlice";
+import type { Token } from "@/types/auth";
 import type { AxiosResponse, InternalAxiosRequestConfig } from "axios";
 import type { AppDispatch, RootState } from "../redux";
-import type { Token } from "@/types/auth";
 
 interface CustomInternalAxiosRequestConfig extends InternalAxiosRequestConfig {
   _isRetry?: boolean;
@@ -27,16 +27,15 @@ export const setupAxiosInterceptors = (dispatch: AppDispatch, getState: () => Ro
       const originalRequest: CustomInternalAxiosRequestConfig | undefined = error.config;
       if (error.status === 401 && originalRequest && !originalRequest._isRetry) {
         originalRequest._isRetry = true;
-        try {
-          const refreshToken = localStorage.getItem(REFRESH_TOKEN);
-          const res = await axios.post<Token>(`${API_URL}/auth/refresh`, { refreshToken });
+        const refreshToken = localStorage.getItem(REFRESH_TOKEN);
+        const res = await axios.post<Token>(`${API_URL}/auth/refresh`, { refreshToken });
 
-          localStorage.setItem(REFRESH_TOKEN, res.data.refreshToken);
-          dispatch(authSlice.actions.setAccessToken(res.data.accessToken));
+        localStorage.setItem(REFRESH_TOKEN, res.data.refreshToken);
+        dispatch(authSlice.actions.setAccessToken(res.data.accessToken));
 
-          return instance.request(originalRequest);
-        } catch (err) {}
+        return instance.request(originalRequest);
       }
+      localStorage.removeItem(REFRESH_TOKEN);
       throw error;
     }
   );
