+++ src/components/UserRightsProtectedRoutes/UserRightsProtectedRoutes.tsx
@@ -2,20 +2,14 @@ import { ForbiddenPage } from "@/pages/ForbiddenPage/ForbiddenPage";
 import { useAppSelector } from "@/redux";
 import { selectProfileRequestData } from "@/selectors.ts/profileSelectors";
 import { isAdminOrModeratorRole } from "@/utils/isAdminOrModeratorRole";
-import LoadingOutlined from "@ant-design/icons/lib/icons/LoadingOutlined";
-import Flex from "antd/es/flex";
-import Spin from "antd/es/spin";
 import { Outlet } from "react-router-dom";
+import { Spinner } from "../Spinner/Spinner";
 
 export const UserRightsProtectedRoutes = () => {
   const { data, status } = useAppSelector(selectProfileRequestData);
 
   if (status.isPending) {
-    return (
-      <Flex justify="center">
-        <Spin size="large" indicator={<LoadingOutlined spin />} />
-      </Flex>
-    );
+    return <Spinner />;
   }
 
   if (isAdminOrModeratorRole(data?.roles)) {
