+++ src/components/constants/todos.ts
@@ -1,3 +1,3 @@
 export const MIN_TODOS_SYMBOLS_COUNT = 2;
 export const MAX_TODOS_SYMBOLS_COUNT = 64;
-export const TODOS_UPDATE_TIME = 50_000;
+export const TODOS_UPDATE_TIME = 5_000;
