+++ src/selectors.ts/authSelectors.ts
@@ -1,6 +1,6 @@
 import { getAsyncRequestData, type AsyncParticle } from "@/utils";
 import { createSelector } from "reselect";
-import type { InitialStateData } from "@/pages/AuthPage/AuthSlice";
+import type { InitialStateData } from "../slices/AuthSlice";
 import type { RootState } from "@/redux";
 
 const selectAuthState = (state: RootState): AsyncParticle<InitialStateData> => state.auth;
