+++ src/slices/AuthSlice.tsx
@@ -1,11 +1,11 @@
+import axios from "axios";
 import { authApi } from "@/api/authApi";
 import { API_URL } from "@/api/instanceApi";
 import { REFRESH_TOKEN } from "@/components/constants/localStorageValues";
 import { addAsyncBuilderCases, getAsyncDataStatus, initAsyncParticle } from "@/utils";
 import { getErrorMessage } from "@/utils/getErrorMessage";
 import { createSlice } from "@reduxjs/toolkit";
-import axios from "axios";
-import { createAppAsyncThunk } from "../../redux";
+import { createAppAsyncThunk } from "@/redux";
 import type { AuthData, Token, UserRegistration } from "@/types/auth";
 import type { AsyncParticle, SliceThunk } from "@/utils";
 import type { PayloadAction } from "@reduxjs/toolkit";
@@ -116,6 +116,9 @@ export const checkAuth: SliceThunk<InitialStateData> = createAppAsyncThunk<Initi
     } catch (err) {
       dispatch(authSlice.actions.setAccessToken(""));
       dispatch(authSlice.actions.setIsAuthorizationFalse());
+      if (axios.isAxiosError(err) && err.status === 401) {
+        return rejectWithValue(null);
+      }
       return rejectWithValue(getErrorMessage(err));
     }
   }
