+++ src/store.ts
@@ -1,9 +1,9 @@
 import { combineSlices, configureStore } from "@reduxjs/toolkit/react";
 import { authApi } from "@/api/authApi";
-import { authSlice } from "@/pages/AuthPage/AuthSlice";
+import { authSlice } from "./slices/AuthSlice";
 import { setupAxiosInterceptors } from "@/api/instanceApi";
 import { profileApi } from "@/api/profileApi";
-import { profileSlice } from "@/pages/ProfilePage/profileSlice";
+import { profileSlice } from "./slices/profileSlice";
 import { usersApi } from "@/api/usersApi";
 
 export const extraArgument = {
